weather-service/requirements.txt
fastapi==0.103.0
uvicorn==0.22.0
pydantic==2.6.1
redis==4.5.5
httpx==0.24.1
python-dateutil==2.8.2
pyjwt[crypto]==2.7.0